requirements.txt
click==7.1.2
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
tqdm==4.47.0
